+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Change Log
+## [Unreleased]
+### Changed
+- Aligned detection of modified monitor configuration closer to Shinobi's behaviour. 
+
 ## 3.0.1
 ### Added
 - More checks to detect invalid monitor configuration, which can lead to internal errors in Shinobi.